Store: MailChimp refactor dasboard and add list name tooltip.
